// src/app/domain/credit-retail/model/credit-application.ts
/**
 * miles-credit-retail
 * This functional service supports the creation and follow-up of **credit applications**.    For more information, check the **Read Me** tab.
 *
 * The version of the OpenAPI document: 1.29.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DynamicAttributeValue } from './dynamic-attribute-value';
import { CustomerReference } from './customer-reference';
import { SalesQuoteReference } from './sales-quote-reference';
import { MultiLanguageValue } from './multi-language-value';
import { EnumValue } from './enum-value';
import { BrokerContactReference } from './broker-contact-reference';
import { ApplicantCreditScoreReference } from './applicant-credit-score-reference';
import { AmendmentQuoteReference } from './amendment-quote-reference';


export interface CreditApplication { 
    /**
     * The ID of the credit application
     */
    creditApplicationId?: string;
    amendmentQuote?: AmendmentQuoteReference;
    salesQuote?: SalesQuoteReference;
    brokerContact?: BrokerContactReference;
    status?: EnumValue;
    decision?: EnumValue;
    /**
     * The date on which the credit application expires, before the credit score decision has been retrieved
     */
    expiryDate?: string;
    /**
     * The date on which the credit application expires, after the credit score decision has been retrieved
     */
    creditDecisionExpiryDate?: string;
    /**
     * The external comment of the credit application
     */
    externalComment?: string;
    vehicleDescription?: MultiLanguageValue;
    mainCustomer?: CustomerReference;
    /**
     * The list of applicant credit scores linked to the credit application
     */
    applicantCreditScores?: Array<ApplicantCreditScoreReference>;
    /**
     * The dynamic attributes of the credit application
     */
    dynamicAttributes?: { [key: string]: DynamicAttributeValue; };
}